Move createMuiTheme to @mui/material
