+++ packages/mui-md-theme/src/createTheme.d.ts
@@ -42,12 +42,6 @@ export interface Theme extends Omit<BaseTheme, 'unstable_sx'> {
 }
 
 /**
- * @deprecated
- * Use `import { createTheme } from '@mui/material/styles'` instead.
- */
-export function createMuiTheme(options?: ThemeOptions, ...args: object[]): Theme;
-
-/**
  * Generate a theme base on the options received.
  * @param options Takes an incomplete theme object and adds the missing parts.
  * @param args Deep merge the arguments with the about to be returned theme.
