+++ packages/mui-md-theme/src/createTheme.js
@@ -121,23 +121,4 @@ function createTheme(options = {}, ...args) {
   return muiTheme;
 }
 
-let warnedOnce = false;
-
-export function createMuiTheme(...args) {
-  if (process.env.NODE_ENV !== 'production') {
-    if (!warnedOnce) {
-      warnedOnce = true;
-      console.error(
-        [
-          'MUI: the createMuiTheme function was renamed to createTheme.',
-          '',
-          "You should use `import { createTheme } from '@mui/material/styles'`",
-        ].join('\n'),
-      );
-    }
-  }
-
-  return createTheme(...args);
-}
-
 export default createTheme;
